backend/requirements.txt
 fastapi 
 uvicorn 
 transformers 
 Pillow 
 pyserial
 torch
 python-multipart
scikit-image

